Addressing Theo's comments part 2
