+++ src/enrichment/models/btsbot.rs
@@ -41,7 +41,7 @@ impl BtsBotModel {
     pub fn get_metadata(
         &self,
         alerts: &[Document],
-        alert_properties: &[Document],
+        alert_properties: &[&Document],
     ) -> Result<Array<f32, Dim<[usize; 2]>>, ModelError> {
         let mut features_batch: Vec<f32> = Vec::with_capacity(alerts.len() * 25);
